# swagger.yaml
openapi: 3.0.0
info:
  description: "This is the swagger documentation and API for the GeoCode project implemented by Peak Performers for the client [5DT](https://5dt.com/)."
  version: "1.0.0"
  title: "Swagger GeoCode"
  termsOfService: "https://swagger.io/terms/"
  contact:
    email: "peakperformers@geocodeapp.tech"
  license:
    name: "Apache 2.0"
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: 'http://geocodeapp.tech:8080/api'
...
paths:
  /User/swapCollectable:
    post:
      tags:
        - "User"
      summary: "Swaps the Collectable the user is currently holding with the given Collectable in the given GeoCode"
      description: "Swap held Collectable"
      operationId: "swapCollectable"
      requestBody:
        required: true
        description: "Request to swap the held Collectable"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwapCollectableRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/SwapCollectableRequest'
      responses:
        '200':
          description: "The user's current Collectable was successfully swapped"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwapCollectableResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/SwapCollectableResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /User/getCurrentCollectable:
    post:
      tags:
        - "User"
      summary: "Get the Collectable the User is currently holding"
      description: "Get the user's current Collectable"
      operationId: "getCurrentCollectable"
      requestBody:
        required: true
        description: "Request to get the user's current Collectable"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetCurrentCollectableRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetCurrentCollectableRequest'
      responses:
        '200':
          description: "Create GeoCode Response"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCurrentCollectableResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetCurrentCollectableResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
...
  /User/getFoundCollectables:
    post:
      tags:
        - "User"
      summary: "Gets the Collectables that the user has ever found"
      description: "Get a user's found Collectables"
      operationId: "getFoundCollectables"
      requestBody:
        required: true
        description: "Request to get the user's found Collectables"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetFoundCollectablesRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetFoundCollectablesRequest'
      responses:
        '200':
          description: "Successfully returned the user's found collectables"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFoundCollectablesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetFoundCollectablesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /User/getFoundGeoCodes:
    post:
      tags:
        - "User"
      summary: "Gets the GeoCodes that the user has ever found"
      description: "Gets the user's found GeoCodes"
      operationId: "getFoundGeoCodes"
      requestBody:
        required: true
        description: "Request to get the user's found GeoCodes"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetFoundGeoCodesRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetFoundGeoCodesRequest'
      responses:
        '200':
          description: "Successfully returned all owned GeoCodes for the user"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFoundGeoCodesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetFoundGeoCodesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /User/getOwnedGeoCodes:
    post:
      tags:
        - "User"
      summary: "Gets the user's owned GeoCodes"
      description: "Get's the user's owned GeoCodes"
      operationId: "getOwnedGeoCodes"
      requestBody:
        required: true
        description: "Request to get the user's owned GeoCodes"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetOwnedGeoCodesRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetOwnedGeoCodesRequest'
      responses:
        '200':
          description: "Successfully returned all owned GeoCodes for the user"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetOwnedGeoCodesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetOwnedGeoCodesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /User/setAdmin:
    post:
      tags:
        - "User"
      summary: "Sets the given user to be an administrator"
      description: "Sets the given user to be an administrator"
      operationId: "setAdmin"
      requestBody:
        required: true
        description: "Request to set a user to be an administrator"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAdminRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/SetAdminRequest'
      responses:
        '200':
          description: "Successfully set the user to be an administrator"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /User/blockUser:
    post:
      tags:
        - "User"
      summary: "Administrator blocks the given user"
      description: "Block a user"
      operationId: "blockUser"
      requestBody:
        required: true
        description: "Request to block a user"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockUserRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/BlockUserRequest'
      responses:
        '200':
          description: "Successfully blocked the user"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /User/getUsers:
    post:
      tags:
        - "User"
      summary: "Get all of the users in the system"
      description: "Get all of the users"
      operationId: "getUsers"
      requestBody:
        required: true
        description: "Request to get all users in the system"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetUsersRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetUsersRequest'
      responses:
        '200':
          description: "Successfully returned all of the users"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUsersResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetUsersResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
...
  /User/getOwnedGeocodes:
    post:
      tags:
        - "User"
      summary: "Get all geocodes associated with a user"
      description: "Get My Geocodes"
      operationId: "getOwnedGeocodes"
      requestBody:
        required: true
        description: "Request to get geocodes belonging to user"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetOwnedGeoCodesRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetOwnedGeoCodesRequest'
      responses:
        '200':
          description: "Successfully gotten geocodes"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetOwnedGeoCodesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetOwnedGeoCodesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
...
components:
  responses:
    UnauthorizedError:
      description: Invalid JWT token
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Difficulty:
      type: string
      description: "The GeoCode's real world locating difficulty"
      enum:
        - "EASY"
        - "MEDIUM"
        - "DIFFICULTY"
        - "INSANE"
    Rarity:
      type: string
      enum:
        - "COMMON"
        - "UNCOMMON"
        - "RARE"
        - "EPIC"
        - "LEGENDARY"
    GeoCode:
      type: "object"
      required:
        - id
        - difficulty
        - description
        - hints
        - qrCode
        - location
        - available
      properties:
        id:
          type: string
          format: uuid
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        available:
          type: boolean
        description:
          type: string
        hints:
          type: array
          items:
            type: string
        collectables:
          $ref: '#/components/schemas/Collectable'
        trackables:
          type: string
        qrCode:
          type: string
        location:
          type: string
      xml:
        name: "GeoCode"
...
    CollectableType:
      type: "object"
      required:
        - id
        - name
        - image
        - rarity
        - set
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        image:
          type: string
        rarity:
          $ref: "#/components/schemas/Rarity"
        set:
          $ref: "#/components/schemas/CollectableSet"
      xml:
        name: "CollectableType"
    CollectableSet:
      type: "object"
      required:
        - id
        - name
        - description
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
      xml:
        name: "CollectableSet"
    Collectable:
      type: "object"
      required:
        - id
        - type
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/CollectableType"
      xml:
        name: "Collectable"
...
    Point:
      type: "object"
      required: 
        - id
        - amount
        - userId
        - leaderBoardId
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: integer
        userId:
          type: string
          format: uuid
        leaderBoardId:
          type: string
          format: uuid
      xml:
        name: "Point"
...
    User:
      type: "object"
      required:
        - id
        - username
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        points:
          type: array
          items:
            $ref: '#/components/schemas/Point'
        currentCollectable:
          $ref: '#/components/schemas/Collectable'
        foundCollectables:
          type: array
          items:
            $ref: '#/components/schemas/Collectable'
        foundGeocodes:
          type: array
          items:
            $ref: '#/components/schemas/GeoCode'
        ownedGeocodes:
          type: array
          items:
            $ref: '#/components/schemas/GeoCode'
    SwapCollectableRequest:
      type: "object"
      required:
        - userID
        - targetGeoCodeID
        - targetCollectableID
      properties:
        userID:
          type: string
          format: uuid
        targetGeoCodeID:
          type: string
          format: uuid
        targetCollectableID:
          type: string
          format: uuid
      xml:
        name: "SwapCollectableRequest"
    SwapCollectableResponse:
      type: "object"
      properties:
        Collectable:
          $ref: '#/components/schemas/Collectable'
      xml:
        name: "SwapCollectableResponse"
    GetCurrentCollectableRequest:
      type: "object"
      required:
        - userID
      properties:
        userID:
          type: string
          format: uuid
      xml:
        name: "GetCurrentCollectableRequest" 
    GetCurrentCollectableResponse:
      type: "object"
      properties:
        Collectable:
          $ref: '#/components/schemas/Collectable'
      xml:
        name: "GetCurrentCollectableResponse"
...
    GetFoundCollectablesRequest:
      type: "object"
      required:
        - userID
      properties:
        userID:
          type: string
          format: uuid
      xml:
        name: "GetFoundCollectablesRequest"
    GetFoundCollectablesResponse:
      type: "object"
      properties:
        collectables:
          type: array
          items:
            $ref: '#/components/schemas/Collectable'
      xml:
        name: "GetFoundCollectablesResponse"
    GetFoundGeoCodesRequest:
      type: "object"
      required:
        - userID
      properties:
        userID:
          type: string
          format: uuid
      xml:
        name: "GetFoundGeocodesRequest"
    GetFoundGeoCodesResponse:
      type: "object"
      properties:
        geocodes:
          type: array
          items:
            $ref: '#/components/schemas/GeoCode'
      xml:
        name: "GetFoundGeoCodesResponse"
    GetOwnedGeoCodesRequest:
      type: "object"
      required:
        - userID
      properties:
        userID:
          type: string
          format: uuid
      xml:
        name: "GetOwnedGeocodesRequest"
    GetOwnedGeoCodesResponse:
      type: "object"
      properties:
        geocodes:
          type: array
          items:
            $ref: '#/components/schemas/GeoCode'
      xml:
        name: "GetOwnedGeocodesResponse"
    SetAdminRequest:
      type: "object"
      required:
        - userID
        - adminID
      properties:
        userID:
          type: string
          format: uuid
        adminID:
          type: string
          format: uuid
      xml:
        name: "SetAdminRequest"
    BlockUserRequest:
      type: "object"
      required:
        - userID
        - adminID
      properties:
        userID:
          type: string
          format: uuid
        adminID:
          type: string
          format: uuid
      xml:
        name: "BlockUserRequest"
    GetUsersRequest:
      type: "object"
      required:
        - adminID
      properties:
        adminID:
          type: string
          format: uuid
      xml:
        name: "GetUsersRequest"
    GetUsersResponse:
      type: "object"
      required:
        - users
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
      xml:
        name: "GetUsersResponse"